pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "polybot"
version = "0.0.1"
description = "Production-grade Polymarket BTC 5-minute trading bot (CLOB V2, maker-first, price-action driven)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "polybot operator" }]

dependencies = [
    # === Polymarket CLOB V2 ===
    # Pinned exactly. Polymarket just shipped V2; the supply chain is not
    # seasoned. Treat as a new dependency: lockstep with constraints.txt.
    "py-clob-client-v2==1.0.0",
    "web3>=6.15",
    "eth-account>=0.10",

    # === Networking ===
    "websockets>=12.0",
    "aiohttp>=3.9",
    "httpx>=0.26",

    # === BTC data feeds ===
    "ccxt>=4.2",

    # === Numerics & analysis ===
    "numpy>=1.26",
    "pandas>=2.1",
    "polars>=0.20",
    "scipy>=1.11",

    # === State, persistence, observability ===
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "PyYAML>=6.0",
    "sqlalchemy>=2.0",
    "loguru>=0.7",
    "rich>=13.7",
    "plotly>=5.18",

    # === Infra ===
    "python-dotenv>=1.0",
    "typer>=0.9",
    # uvloop is Linux/macOS only; selected via environment marker
    "uvloop>=0.19; sys_platform != 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "hypothesis>=6.92",
    "freezegun>=1.4",
    "ruff>=0.3",
    "black>=24.1",
    "mypy>=1.8",
    "types-PyYAML",
    "pre-commit>=3.6",
]

[project.scripts]
polybot = "polybot.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/polybot"]

# ============================================================================
# Tooling
# ============================================================================

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle / pyflakes
    "I",             # isort
    "B",             # bugbear
    "UP",            # pyupgrade
    "N",             # pep8-naming
    "C4",            # comprehensions
    "SIM",           # simplify
    "RUF",           # ruff
]
ignore = [
    "E501",          # handled by black
    "B008",          # function call in default argument (typer uses this)
]

# NOTE on log-format drift:
# Ruff's flake8-logging-format `G` rules don't fire on loguru calls
# (they're scoped to stdlib `logging`). The drift guard lives in
# tests/unit/test_logger_format_drift.py — a fast regex scan that fails
# CI if any `log.<level>("...%s...", arg)` re-enters the codebase.

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["N802", "N803"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
disallow_untyped_decorators = true
files = ["src/polybot", "config"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "py_clob_client_v2.*",
    "ccxt.*",
    "plotly.*",
    "scipy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "integration: integration tests (touch network or real services)",
    "property: hypothesis property-based tests",
    "slow: tests that take >5 seconds",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
branch = true
source = ["src/polybot"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
